updated the files with changes only for get workspace details and get run details
