+++ cs-hashicorp-terraform/src/main/java/io/cloudslang/content/hashicorp/terraform/entities/GetWorkspaceDetailsInputs.java
@@ -15,16 +15,15 @@
 
 package io.cloudslang.content.hashicorp.terraform.entities;
 
-import io.cloudslang.content.hashicorp.terraform.utils.Inputs;
 import org.jetbrains.annotations.NotNull;
 
 public class GetWorkspaceDetailsInputs {
 
     private final String workspaceName;
-    private final Inputs commonInputs;
+    private final TerraformCommonInputs commonInputs;
 
     @java.beans.ConstructorProperties({"workspaceName", "commonInputs"})
-    public GetWorkspaceDetailsInputs(String workspaceName, Inputs commonInputs) {
+    public GetWorkspaceDetailsInputs(String workspaceName, TerraformCommonInputs commonInputs) {
         this.workspaceName = workspaceName;
         this.commonInputs = commonInputs;
     }
@@ -40,14 +39,14 @@ public class GetWorkspaceDetailsInputs {
     }
 
     @NotNull
-    public Inputs getCommonInputs() {
+    public TerraformCommonInputs getCommonInputs() {
         return this.commonInputs;
     }
 
 
     public static class GetWorkspaceDetailsInputsBuilder {
         private String workspaceName;
-        private Inputs commonInputs;
+        private TerraformCommonInputs commonInputs;
 
         GetWorkspaceDetailsInputsBuilder() {
 
@@ -59,7 +58,7 @@ public class GetWorkspaceDetailsInputs {
             return this;
         }
 
-               public GetWorkspaceDetailsInputs.GetWorkspaceDetailsInputsBuilder commonInputs(@NotNull final Inputs commonInputs) {
+               public GetWorkspaceDetailsInputs.GetWorkspaceDetailsInputsBuilder commonInputs(@NotNull final TerraformCommonInputs commonInputs) {
             this.commonInputs = commonInputs;
             return this;
         }
