+++ cs-hashicorp-terraform/src/main/java/io/cloudslang/content/hashicorp/terraform/services/WorkspaceImpl.java
@@ -14,12 +14,8 @@
  */
 package io.cloudslang.content.hashicorp.terraform.services;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.cloudslang.content.hashicorp.terraform.entities.CreateWorkspaceInputs;
 import io.cloudslang.content.hashicorp.terraform.entities.GetWorkspaceDetailsInputs;
-import io.cloudslang.content.hashicorp.terraform.services.CreateWorkspaceModels.CreateWorkspaceBody;
-import io.cloudslang.content.hashicorp.terraform.utils.Inputs;
+import io.cloudslang.content.hashicorp.terraform.entities.TerraformCommonInputs;
 import io.cloudslang.content.httpclient.entities.HttpClientInputs;
 import io.cloudslang.content.httpclient.services.HttpClientService;
 import org.apache.http.client.utils.URIBuilder;
@@ -27,38 +23,16 @@ import org.jetbrains.annotations.NotNull;
 import java.util.Map;
 import static io.cloudslang.content.hashicorp.terraform.services.HttpCommons.setCommonHttpInputs;
 import static io.cloudslang.content.hashicorp.terraform.utils.Constants.Common.ANONYMOUS;
-import static io.cloudslang.content.hashicorp.terraform.utils.Constants.CreateWorkspace.WORKSPACE_PATH;
-import static io.cloudslang.content.hashicorp.terraform.utils.Constants.CreateWorkspace.WORKSPACE_TYPE;
 import static io.cloudslang.content.hashicorp.terraform.utils.Constants.GetWorkspaceDetails.GET_WORKSPACE_PATH;
 import static io.cloudslang.content.hashicorp.terraform.utils.HttpUtils.*;
 import static io.cloudslang.content.hashicorp.terraform.utils.Constants.Common.*;
-import static org.apache.commons.lang3.StringUtils.EMPTY;
 
 public class WorkspaceImpl {
 
     @NotNull
-    public static Map<String, String> createWorkspace(@NotNull final CreateWorkspaceInputs createWorkspaceInputs) throws Exception {
-        final HttpClientInputs httpClientInputs = new HttpClientInputs();
-        final Inputs commonInputs = createWorkspaceInputs.getCommonInputs();
-        httpClientInputs.setUrl(createWorkspaceUrl(createWorkspaceInputs.getCommonInputs().getOrganizationName()));
-        setCommonHttpInputs(httpClientInputs, commonInputs);
-        httpClientInputs.setAuthType(ANONYMOUS);
-        httpClientInputs.setMethod(POST);
-        httpClientInputs.setContentType(APPLICATION_VND_API_JSON);
-        if (commonInputs.getRequestBody().equals(EMPTY)) {
-            httpClientInputs.setBody(createWorkspaceBody(createWorkspaceInputs));
-        } else {
-            httpClientInputs.setBody(commonInputs.getRequestBody());
-        }
-        httpClientInputs.setResponseCharacterSet(commonInputs.getResponseCharacterSet());
-        httpClientInputs.setHeaders(getAuthHeaders(commonInputs.getAuthToken()));
-        return new HttpClientService().execute(httpClientInputs);
-    }
-
-    @NotNull
     public static Map<String, String> getWorkspaceDetails(@NotNull final GetWorkspaceDetailsInputs getWorkspaceDetailsInputs) throws Exception {
         final HttpClientInputs httpClientInputs = new HttpClientInputs();
-        final Inputs commonInputs = getWorkspaceDetailsInputs.getCommonInputs();
+        final TerraformCommonInputs commonInputs = getWorkspaceDetailsInputs.getCommonInputs();
         httpClientInputs.setUrl(getWorkspaceDetailsUrl(getWorkspaceDetailsInputs.getCommonInputs().getOrganizationName(),getWorkspaceDetailsInputs.getWorkspaceName()));
         httpClientInputs.setAuthType(ANONYMOUS);
         httpClientInputs.setMethod(GET);
@@ -66,13 +40,6 @@ public class WorkspaceImpl {
         httpClientInputs.setContentType(APPLICATION_VND_API_JSON);
         setCommonHttpInputs(httpClientInputs, commonInputs);
         return new HttpClientService().execute(httpClientInputs);
-    }
-
-    @NotNull
-    private static String createWorkspaceUrl(@NotNull final String organizationName) throws Exception {
-        final URIBuilder uriBuilder = getUriBuilder();
-        uriBuilder.setPath(getWorkspacePath(organizationName));
-        return uriBuilder.build().toURL().toString();
     }
 
     @NotNull
@@ -93,52 +60,5 @@ public class WorkspaceImpl {
                 .append(workspaceName);
         return pathString.toString();
     }
-    @NotNull
-    public static String getWorkspacePath(@NotNull final String organizationName) {
-        StringBuilder pathString = new StringBuilder()
-                .append(API)
-                .append(API_VERSION)
-                .append(ORGANIZATION_PATH)
-                .append(organizationName)
-                .append(WORKSPACE_PATH);
-        return pathString.toString();
-    }
 
-    @NotNull
-    public static String createWorkspaceBody(CreateWorkspaceInputs createWorkspaceInputs) {
-        ObjectMapper createWorkspaceMapper = new ObjectMapper();
-        CreateWorkspaceBody createBody = new CreateWorkspaceBody();
-        CreateWorkspaceBody.CreateWorkspaceData createWorkspaceData = createBody.new CreateWorkspaceData();
-        CreateWorkspaceBody.Attributes attributes = createBody.new Attributes();
-        String requestBody = EMPTY;
-
-        attributes.setName(createWorkspaceInputs.getWorkspaceName());
-        attributes.setTerraform_version(createWorkspaceInputs.getCommonInputs().getTerraformVersion());
-        attributes.setDescription(createWorkspaceInputs.getWorkspaceDescription());
-
-        CreateWorkspaceBody.VCSRepo vcsRepo = createBody.new VCSRepo();
-
-        vcsRepo.setIdentifier(createWorkspaceInputs.getVcsRepoId());
-        vcsRepo.setOauthTokenId(createWorkspaceInputs.getOauthTokenId());
-        vcsRepo.setBranch(createWorkspaceInputs.getVcsBranch());
-        vcsRepo.setIsDefaultBranch(createWorkspaceInputs.isDefaultBranch());
-
-        attributes.setVcsRepo(vcsRepo);
-
-        createWorkspaceData.setAttributes(attributes);
-        createWorkspaceData.setType(WORKSPACE_TYPE);
-
-        createBody.setData(createWorkspaceData);
-
-
-        try {
-            requestBody = createWorkspaceMapper.writeValueAsString(createBody);
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-        }
-
-        System.out.println(requestBody);
-        return requestBody;
-
-    }
 }
